Fix task-group importance sorting ..
i.e. so task groups containing importance tasks are moved to the
top of the time group.

In English: tasks are grouped by type, e.g. all the errands are
grouped together. These groups are grouped into time of day (because
some errands have to be done in the morning/afternoon/etc).
If a group of morning-errands contains an urgent item, the whole group is
moved to the top of the morning list.

Next step is to sort errands/chores/etc. by importance within
their groupings. So highest importance errands come first.